+++ GANDLF/training_loop.py
@@ -17,9 +17,10 @@ import numpy as np
 from medcam import medcam
 from GANDLF.logger import Logger
 from GANDLF.models import global_models_dict
+from GANDLF.schedulers import global_schedulers_dict
+from GANDLF.optimizers import global_optimizer_dict
 from GANDLF.parameterParsing import (
-    get_optimizer,
-    get_scheduler,
+    # get_optimizer,
     get_loss_and_metrics,
 )
 from GANDLF.utils import (
@@ -635,7 +636,12 @@ def validate_network(
         )
 
     if scheduler is not None:
-        if params["scheduler"] == "reduce-on-plateau":
+        if params["scheduler"]["type"] in [
+            "reduce_on_plateau",
+            "reduce-on-plateau",
+            "plateau",
+            "reduceonplateau",
+        ]:
             scheduler.step(average_epoch_valid_loss)
         else:
             scheduler.step()
@@ -709,6 +715,7 @@ def training_loop(
         shuffle=True,
         pin_memory=False,  # params["pin_memory_dataloader"], # this is going OOM if True - needs investigation
     )
+    params["training_samples_size"] = len(train_dataloader.dataset)
 
     val_dataloader = DataLoader(
         validation_data_for_torch,
@@ -749,19 +756,21 @@ def training_loop(
         params["weights"], params["class_weights"] = None, None
 
     # Fetch the optimizers
-    optimizer = get_optimizer(
-        optimizer_name=params["opt"],
-        model=model,
-        learning_rate=params["learning_rate"],
-    )
+    params["model_parameters"] = model.parameters()
+    optimizer = global_optimizer_dict[params["optimizer"]["type"]](params)
+    params["optimizer_object"] = optimizer
 
-    scheduler = get_scheduler(
-        which_scheduler=params["scheduler"],
-        optimizer=optimizer,
-        batch_size=params["batch_size"],
-        training_samples_size=len(train_dataloader.dataset),
-        learning_rate=params["learning_rate"],
-    )
+    if not ("step_size" in params["scheduler"]):
+        params["scheduler"]["step_size"] = (
+            params["training_samples_size"] / params["learning_rate"]
+        )
+
+    scheduler = global_schedulers_dict[params["scheduler"]["type"]](params)
+
+    # these keys contain generators, and are not needed beyond this point in params
+    generator_keys_to_remove = ["optimizer_object", "model_parameters"]
+    for key in generator_keys_to_remove:
+        params.pop(key, None)
 
     # Start training time here
     start_time = time.time()